# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CRMContactRequest": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "name"
        ],
        "title": "CRMContactRequest",
        "type": "object"
      },
      "CRMContactResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "CRMContactResponse",
        "type": "object"
      },
...
      "ComparePropertiesRequest": {
        "properties": {
          "property_ids": {
            "items": {
              "type": "string"
            },
            "title": "Property Ids",
            "type": "array"
          }
        },
        "required": [
          "property_ids"
        ],
        "title": "ComparePropertiesRequest",
        "type": "object"
      },
      "ComparePropertiesResponse": {
        "properties": {
          "properties": {
            "items": {
              "$ref": "#/components/schemas/ComparedProperty"
            },
            "title": "Properties",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/CompareSummary"
          }
        },
        "required": [
          "properties",
          "summary"
        ],
        "title": "ComparePropertiesResponse",
        "type": "object"
      },
...
      "DataEnrichmentRequest": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          }
        },
        "required": [
          "address"
        ],
        "title": "DataEnrichmentRequest",
        "type": "object"
      },
      "DataEnrichmentResponse": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "default": {},
            "title": "Data",
            "type": "object"
          }
        },
        "required": [
          "address"
        ],
        "title": "DataEnrichmentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestmentAnalysisResult": {
        "description": "Result from investment property analysis.",
        "properties": {
          "annual_cash_flow": {
            "title": "Annual Cash Flow",
            "type": "number"
          },
          "annual_expenses": {
            "title": "Annual Expenses",
            "type": "number"
          },
          "annual_income": {
            "title": "Annual Income",
            "type": "number"
          },
          "cap_rate": {
            "title": "Cap Rate",
            "type": "number"
          },
          "cash_on_cash_roi": {
            "title": "Cash On Cash Roi",
            "type": "number"
          },
          "gross_yield": {
            "title": "Gross Yield",
            "type": "number"
          },
          "investment_score": {
            "title": "Investment Score",
            "type": "number"
          },
          "monthly_cash_flow": {
            "title": "Monthly Cash Flow",
            "type": "number"
          },
          "monthly_expenses": {
            "title": "Monthly Expenses",
            "type": "number"
          },
          "monthly_income": {
            "title": "Monthly Income",
            "type": "number"
          },
          "monthly_mortgage": {
            "title": "Monthly Mortgage",
            "type": "number"
          },
          "net_yield": {
            "title": "Net Yield",
            "type": "number"
          },
          "score_breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Score Breakdown",
            "type": "object"
          },
          "total_investment": {
            "title": "Total Investment",
            "type": "number"
          }
        },
        "required": [
          "monthly_cash_flow",
          "annual_cash_flow",
          "cash_on_cash_roi",
          "cap_rate",
          "gross_yield",
          "net_yield",
          "total_investment",
          "monthly_income",
          "monthly_expenses",
          "annual_income",
          "annual_expenses",
          "monthly_mortgage",
          "investment_score",
          "score_breakdown"
        ],
        "title": "InvestmentAnalysisResult",
        "type": "object"
      },
...
      "ToolInfo": {
        "description": "Information about an available tool.",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ],
        "title": "ToolInfo",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "AI Real Estate Assistant API V4",
    "title": "AI Real Estate Assistant - Modern",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tools": {
      "get": {
        "description": "List available tools.",
        "operationId": "list_tools_api_v1_tools_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ToolInfo"
                  },
                  "title": "Response List Tools Api V1 Tools Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Tools",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/compare-properties": {
      "post": {
        "operationId": "compare_properties_api_v1_tools_compare_properties_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparePropertiesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparePropertiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Compare Properties",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/crm-sync-contact": {
      "post": {
        "operationId": "crm_sync_contact_api_v1_tools_crm_sync_contact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CRMContactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CRMContactResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Crm Sync Contact",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/enrich-address": {
      "post": {
        "operationId": "enrich_address_api_v1_tools_enrich_address_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataEnrichmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnrichmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Enrich Address",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/investment-analysis": {
      "post": {
        "description": "Calculate investment property metrics including ROI, cap rate, cash flow, and rental yield.",
        "operationId": "calculate_investment_analysis_api_v1_tools_investment_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestmentAnalysisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentAnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Calculate Investment Analysis",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/legal-check": {
      "post": {
        "operationId": "legal_check_api_v1_tools_legal_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalCheckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Legal Check",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/location-analysis": {
      "post": {
        "operationId": "location_analysis_api_v1_tools_location_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Location Analysis",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/mortgage-calculator": {
      "post": {
        "description": "Calculate mortgage payments.",
        "operationId": "calculate_mortgage_api_v1_tools_mortgage_calculator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MortgageInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MortgageResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Calculate Mortgage",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/neighborhood-quality": {
      "post": {
        "description": "Calculate neighborhood quality index including safety, schools, amenities, walkability, and green space.",
        "operationId": "neighborhood_quality_api_v1_tools_neighborhood_quality_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NeighborhoodQualityInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborhoodQualityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Neighborhood Quality",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/price-analysis": {
      "post": {
        "operationId": "price_analysis_api_v1_tools_price_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Price Analysis",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/tco-calculator": {
      "post": {
        "description": "Calculate Total Cost of Ownership for a property.\nIncludes mortgage, property taxes, insurance, HOA fees, utilities, maintenance, and parking.",
        "operationId": "calculate_tco_api_v1_tools_tco_calculator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TCOInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TCOResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Calculate Tco",
        "tags": [
          "Tools"
        ]
      }
    },
    "/api/v1/tools/valuation": {
      "post": {
        "operationId": "valuation_api_v1_tools_valuation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValuationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Valuation",
        "tags": [
          "Tools"
        ]
      }
    },
...
  }
}